User 64 bit params mask.
ifacename put is going to put us over the limit, so have
the session param mask match the kernel one.

+++ usr/initiator.h
@@ -240,7 +240,7 @@ typedef struct iscsi_session {
 	int password_in_length;
 	iscsi_conn_t conn[ISCSI_CONN_MAX];
 	int ctrl_fd;
-	uint32_t param_mask;
+	uint64_t param_mask;
 
 	/* connection reopens during recovery */
 	int reopen_cnt;
